feat: 分组代理密钥优化
